backend/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="CardioPulse AI Backend", version="1.0.0")

# FDA Compliance: Backend handles API Keys securely
genai.configure(apiKey=os.getenv("GEMINI_API_KEY"))
model = genai.GenerativeModel('gemini-1.5-flash')

class AnalysisRequest(BaseModel):
    image_data: str  # Base64 encoded image

@app.get("/api/health")
async def health_check():
    return {"status": "operational", "system": "CardioPulse AI"}

@app.post("/api/analyze/ecg")
async def analyze_ecg(request: AnalysisRequest):
    try:
        # Implementation of ECG analysis logic from geminiService.ts
        # This keeps sensitive diagnostic logic on the server side
        # for FDA/HIPAA compliance patterns
        prompt = "Analyze this ECG waveform..."
        # (Real implementation would go here)
        return {"result": "Analysis processed on Python backend"}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
